fix #12
